# packages/api-client/openapi.yaml
openapi: 3.1.0
info:
  title: Arc - みらい予報 API
  version: 0.2.0
  description: 予約・決済・相談員管理 API
servers:
  - url: /api
paths:
  /customer/me:
    get:
      operationId: getCustomerMe
      summary: 顧客プロフィール取得
      tags: [customer]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "404":
          description: 未登録（サインアップ前）
    patch:
      operationId: updateCustomerProfile
      summary: 顧客プロフィール更新
      tags: [customer]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, birthDate]
              properties:
                displayName:
                  type: string
                primaryEmail:
                  type: string
                birthDate:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
    delete:
      operationId: withdrawCustomer
      summary: 顧客退会
      tags: [customer]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /customer/me/signup:
    post:
      operationId: signupCustomer
      summary: 顧客会員登録（匿名 / Google 連携）
      tags: [customer]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, birthDate]
              properties:
                displayName:
                  type: string
                birthDate:
                  type: string
                primaryEmail:
                  type: string
                providerUid:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [userId, isNew]
                properties:
                  userId:
                    type: string
                  isNew:
                    type: boolean
  /customer/me/coupons:
    get:
      operationId: getCustomerCoupons
      summary: 顧客の受け取り済みクーポン一覧
      tags: [customer]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [coupons]
                properties:
                  coupons:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomerCoupon"
...
  /auth/zoom/authorize:
    post:
      operationId: createZoomAuthorizeUrl
      summary: Zoom 連携の認可 URL を発行
      tags: [customer]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
  /auth/zoom/revoke:
    post:
      operationId: revokeZoomConnection
      summary: Zoom 連携を解除
      tags: [customer]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
  /organizations/{organizationId}/bookings:
    post:
      operationId: createBooking
      summary: 予約作成
      tags: [booking]
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBookingResponse"
  /organizations/{organizationId}/bookings/{bookingId}/cancel:
    post:
      operationId: cancelBooking
      summary: 予約キャンセル
      tags: [booking]
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/BookingIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cancelledBy]
              properties:
                cancelledBy:
                  type: string
                  enum: [customer, admin]
                token:
                  type: string
                  description: "顧客キャンセル時に必須。形式: bookingId.expiresAt.signature"
      responses:
        "200":
          description: OK
  /organizations/{organizationId}/bookings/{bookingId}/setup-payment:
    post:
      operationId: setupPayment
      summary: 決済セットアップ
      tags: [booking]
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/BookingIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupPaymentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupPaymentResponse"
  /organizations/{organizationId}/bookings/{bookingId}/charge:
    post:
      operationId: chargePayment
      summary: 手動課金
      tags: [booking]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/BookingIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [method]
              properties:
                method:
                  type: string
                  enum: [manual]
      responses:
        "200":
          description: OK
...
components:
  parameters:
    OrganizationIdParam:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
    BookingIdParam:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    CustomerProfile:
      type: object
      required:
        [
          userId,
          authUid,
          displayName,
          primaryEmail,
          birthDate,
          status,
          authProviders,
          hasActiveZoomConnection,
          zoomEmail,
          createdAt,
          updatedAt,
        ]
      properties:
        userId:
          type: string
        authUid:
          type: string
        displayName:
          type: string
        primaryEmail:
          type: string
          nullable: true
        birthDate:
          type: string
        status:
          type: string
          enum: [active, withdrawn]
        authProviders:
          type: array
          items:
            $ref: "#/components/schemas/CustomerAuthProviderLink"
        hasActiveZoomConnection:
          type: boolean
        zoomEmail:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
    CustomerAuthProviderLink:
      type: object
      required: [providerId, linkedAt]
      properties:
        providerId:
          type: string
        linkedAt:
          type: string
    CustomerCoupon:
      type: object
      required:
        [
          userCouponId,
          couponId,
          organizationId,
          couponName,
          amountJPY,
          type,
          receivedAt,
          expiresAt,
          redeemedAt,
          isRedeemable,
        ]
      properties:
        userCouponId:
          type: string
        couponId:
          type: string
        organizationId:
          type: string
        couponName:
          type: string
        amountJPY:
          type: integer
          minimum: 0
        type:
          $ref: "#/components/schemas/CouponType"
        receivedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        redeemedAt:
          type: string
          format: date-time
          nullable: true
        isRedeemable:
          type: boolean
...
    CreateBookingRequest:
      type: object
      required:
        [
          customerName,
          customerEmail,
          customerPhone,
          customerBirthDate,
          selectionId,
        ]
      properties:
        slotId:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        customerPhone:
          type: string
        customerBirthDate:
          type: string
          format: date
        consultantContent:
          type: string
        selectionId:
          type: string
        selectedUserCouponId:
          type: string
          description: 適用する UserCoupon の ID（1枚のみ、任意）
...
    CreateBookingResponse:
      type: object
      required: [bookingId, joinUrl, bookingActionToken]
      properties:
        bookingId:
          type: string
        joinUrl:
          type: string
          format: uri
        bookingActionToken:
          type: string
    SetupPaymentRequest:
      type: object
      required: [paymentMethodType, bookingActionToken]
      properties:
        paymentMethodType:
          type: string
          enum: [card, paypay]
        bookingActionToken:
          type: string
    SetupPaymentResponse:
      type: object
      required: [customerSecret, mode]
      properties:
        customerSecret:
          type: string
        mode:
          type: string
          enum: [setup, payment]
...
    CouponType:
      type: string
      enum: [welcome, birthday]
...